+++ CTF/src/main/java/ca/mcgill/science/ctf/fragments/PreTicketFragment.java
@@ -1,5 +1,6 @@
 package ca.mcgill.science.ctf.fragments;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.CallSuper;
 import android.support.annotation.Nullable;
@@ -122,8 +123,8 @@ public class PreTicketFragment extends Fragment implements CFragmentCore {
     }
 
     @Override
-    public void onResume() {
-        super.onResume();
+    public void onAttach(Context context) {
+        super.onAttach(context);
         EventBus.getDefault().post(new CFabEvent(GoogleMaterial.Icon.gmd_send, v -> {
             ((MainActivity) getContext()).addFragment(TicketFragment.getInstance(user, problemText.getText().toString(), printer));
         }));